clean up bug fix

+++ bench/syclblas/syclblas_benchmark_level2.cpp
@@ -31,8 +31,9 @@ using namespace blas;
 
 BENCHMARK_NAME_FORMAT(syclblas_level_2) {
   std::ostringstream fname;
-  fname << benchmark<>::typestr<ElemT>() << "_" << name() << "_" << std::get<0>(params)
-        << "_" << std::get<1>(params) << "_" << std::get<2>(params);
+  fname << benchmark<>::typestr<ElemT>() << "_" << name() << "_"
+        << std::get<0>(params) << "_" << std::get<1>(params) << "_"
+        << std::get<2>(params);
   return fname.str();
 }
 
@@ -47,7 +48,8 @@ BENCHMARK(gemv, syclblas_level_2) {
   IndexType vlen = t_str[0] == 'n' ? n : m;
   IndexType rlen = t_str[0] == 'n' ? m : n;
 
-  size_t n_fl_ops = m * n * 2;
+  size_t n_fl_ops =
+      static_cast<size_t>(m) * static_cast<size_t>(n) * static_cast<size_t>(2);
 
   IndexType lda = m;
   long incX = 1;
